small refacto: mutualize entities like Name, Location, Persons in a ts
